# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Templarc",
...
    "contact": {
      "name": "Templarc"
    },
    "license": {
      "name": "Private"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/templates/{template_id}/resolve-params": {
      "get": {
        "tags": [
          "Render",
          "Render"
        ],
        "summary": "Resolve parameters for form rendering",
        "description": "Returns the full enriched parameter set for a template, ready for the UI to build a dynamic input form.  Runs all `on_load` data sources in parallel and merges enrichments (prefill values, dropdown options, readonly overrides) into each parameter descriptor.",
        "operationId": "resolve_params_templates__template_id__resolve_params_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "template_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Template Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormDefinitionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/templates/{template_id}/render": {
      "post": {
        "tags": [
          "Render",
          "Render"
        ],
        "summary": "Render a template",
        "description": "Render a template with provided parameter values.  Pass `?persist=false` to skip writing to `render_history` (useful for quick previews or CI dry-runs).  The response `output` includes the metadata header prepended to the rendered body.",
        "operationId": "render_template_templates__template_id__render_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "template_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Template Id"
            }
          },
          {
            "name": "persist",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Set false to skip render_history storage",
              "default": true,
              "title": "Persist"
            },
            "description": "Set false to skip render_history storage"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenderRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RenderOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/templates/{template_id}/on-change/{param_name}": {
      "post": {
        "tags": [
          "Render",
          "Render"
        ],
        "summary": "Re-resolve datasources after a parameter change",
        "description": "Call this when the user edits a form field.  Returns updated enrichments (prefill values, dropdown options) for any parameters whose datasources are triggered by the changed param.  Cascading triggers are resolved automatically with loop detection.\n\nResponse is a mapping of parameter name \u2192 enrichment object, e.g.:\n```json\n{\"router.site_id\": {\"prefill\": \"42\", \"readonly\": true, \"source_id\": \"netbox\"}}\n```",
        "operationId": "on_change_templates__template_id__on_change__param_name__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "template_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Template Id"
            }
          },
          {
            "name": "param_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Param Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OnChangeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response On Change Templates  Template Id  On Change  Param Name  Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/render-history": {
      "get": {
        "tags": [
          "Render",
          "Render History"
        ],
        "summary": "List render history",
        "description": "Return render history records, newest first, with optional filters.",
        "operationId": "list_render_history_render_history_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "template_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Template Id"
            }
          },
          {
            "name": "date_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date From"
            }
          },
          {
            "name": "date_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date To"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RenderHistoryListOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/render-history/{history_id}": {
      "get": {
        "tags": [
          "Render",
          "Render History"
        ],
        "summary": "Get a single render history record",
        "description": "Fetch one render history record by ID. Includes the full resolved parameter set and the raw rendered output. Use this to inspect what was rendered and replay it via the re-render endpoint.",
        "operationId": "get_render_history_render_history__history_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "history_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "History Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RenderHistoryOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/render-history/{history_id}/re-render": {
      "post": {
        "tags": [
          "Render",
          "Render History"
        ],
        "summary": "Re-render using stored parameters",
        "description": "Re-runs a previous render using the exact same resolved parameters.  The template body is re-loaded from Git, so changes to the .j2 file are reflected in the output.  Pass `template_id` in the body to apply the stored params to a different template.",
        "operationId": "re_render_render_history__history_id__re_render_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "history_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "History Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReRenderRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RenderOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EnrichedParameterOut": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "scope": {
            "type": "string",
            "title": "Scope"
          },
          "widget_type": {
            "type": "string",
            "title": "Widget Type"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "help_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Help Text"
          },
          "default_value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Value"
          },
          "required": {
            "type": "boolean",
            "title": "Required"
          },
          "sort_order": {
            "type": "integer",
            "title": "Sort Order"
          },
          "is_derived": {
            "type": "boolean",
            "title": "Is Derived"
          },
          "prefill": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Prefill"
          },
          "options": {
            "items": {
              "type": "object"
            },
            "type": "array",
            "title": "Options"
          },
          "readonly": {
            "type": "boolean",
            "title": "Readonly",
            "default": false
          },
          "source_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Id"
          }
        },
        "type": "object",
        "required": [
          "name",
          "scope",
          "widget_type",
          "label",
          "description",
          "help_text",
          "default_value",
          "required",
          "sort_order",
          "is_derived"
        ],
        "title": "EnrichedParameterOut",
        "description": "A resolved parameter enriched with any data-source prefills/options."
      },
...
      "FormDefinitionOut": {
        "properties": {
          "template_id": {
            "type": "integer",
            "title": "Template Id"
          },
          "parameters": {
            "items": {
              "$ref": "#/components/schemas/EnrichedParameterOut"
            },
            "type": "array",
            "title": "Parameters"
          },
          "inheritance_chain": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Inheritance Chain"
          }
        },
        "type": "object",
        "required": [
          "template_id",
          "parameters",
          "inheritance_chain"
        ],
        "title": "FormDefinitionOut",
        "description": "Response from GET /templates/{id}/resolve-params."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OnChangeRequest": {
        "properties": {
          "current_params": {
            "type": "object",
            "title": "Current Params",
            "description": "Current form state \u2014 all param values known so far"
          }
        },
        "type": "object",
        "title": "OnChangeRequest",
        "description": "Body for POST /templates/{id}/on-change/{param_name}."
      },
...
      "ReRenderRequest": {
        "properties": {
          "template_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Template Id",
            "description": "Override the original template. Useful for testing the same params against a new template version. Defaults to the original template."
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "persist": {
            "type": "boolean",
            "title": "Persist",
            "default": true
          }
        },
        "type": "object",
        "title": "ReRenderRequest",
        "description": "Body for POST /render-history/{id}/re-render."
      },
      "RenderHistoryListOut": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RenderHistoryOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "RenderHistoryListOut",
        "description": "Paginated list of render history records."
      },
      "RenderHistoryOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "template_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Template Id"
          },
          "template_git_sha": {
            "type": "string",
            "title": "Template Git Sha"
          },
          "resolved_parameters": {
            "type": "object",
            "title": "Resolved Parameters"
          },
          "raw_output": {
            "type": "string",
            "title": "Raw Output"
          },
          "rendered_by": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rendered By"
          },
          "rendered_at": {
            "type": "string",
            "format": "date-time",
            "title": "Rendered At"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "id",
          "template_id",
          "template_git_sha",
          "resolved_parameters",
          "raw_output",
          "rendered_by",
          "rendered_at",
          "notes"
        ],
        "title": "RenderHistoryOut",
        "description": "A render history record."
      },
      "RenderOut": {
        "properties": {
          "output": {
            "type": "string",
            "title": "Output",
            "description": "Full rendered output including the metadata header"
          },
          "render_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Render Id",
            "description": "ID of the stored RenderHistory record; None when persist=false"
          },
          "template_id": {
            "type": "integer",
            "title": "Template Id"
          },
          "git_sha": {
            "type": "string",
            "title": "Git Sha"
          }
        },
        "type": "object",
        "required": [
          "output",
          "template_id",
          "git_sha"
        ],
        "title": "RenderOut",
        "description": "Response from POST /templates/{id}/render.",
        "examples": [
          {
            "git_sha": "a1b2c3d4e5f6789012345678901234567890abcd",
            "output": "! Rendered by admin at 2024-01-15T10:30:00Z\n! Template: cisco_891\nhostname edge-01.dc.company.com\n",
            "render_id": 42,
            "template_id": 5
          }
        ]
      },
      "RenderRequest": {
        "properties": {
          "params": {
            "type": "object",
            "title": "Params",
            "description": "Template-local parameter values keyed by flat dot-notation name (e.g. 'router.hostname'). glob.* and proj.* keys are ignored."
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes",
            "description": "Optional freeform render notes stored in history"
          }
        },
        "type": "object",
        "title": "RenderRequest",
        "description": "Body for POST /templates/{id}/render.",
        "examples": [
          {
            "notes": "Production render for edge-01",
            "params": {
              "router.hostname": "edge-01.dc.company.com",
              "router.loopback": "10.0.0.1",
              "router.site_id": "42"
            }
          }
        ]
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}